Module Make.Path

The type of Path.t allows to specify a rank-one sub-tensor out of a rank-n tensor.

type 'a t

The type of paths. A rank-n tensor is a product of lower rank tensors, so a path amounts to a sequence of left-right projects down to a rank-one tensor.

val empty : pos t

The empty path.

val l : 'a t -> ('a * 'b) t

Select the left sub-tensor.

val r : 'a t -> ('b * 'a) t

Select the right sub-tensor.

val left : unit -> (pos * 'a) t

left () evaluates to l empty.

val right : unit -> ('a * pos) t

right () evaluates to l empty.